move checks into methods

+++ inc/classes/Argument.php
@@ -25,11 +25,58 @@ class Argument extends Relation {
 
 
 	/**
+	 * arguments older than this may not be edited anymore
+	 *
+	 * @return integer
+	 */
+	static function edit_limit() {
+		static $edit_limit = false;
+		if ($edit_limit===false) $edit_limit = strtotime("- ".ARGUMENT_EDIT_INTERVAL);
+		return $edit_limit;
+	}
+
+
+	/**
+	 * wrapper for create()
+	 */
+	function add() {
+		if (mb_strlen($this->title) > self::title_length) {
+			$this->title = limitstr($this->title, self::title_length);
+			warning(sprintf(_("The title has been truncated to the maximum allowed length of %d characters!"), self::title_length));
+		}
+		if (mb_strlen($this->content) > self::content_length) {
+			$this->content = limitstr($this->content, self::content_length);
+			warning(sprintf(_("The content has been truncated to the maximum allowed length of %d characters!"), self::content_length));
+		}
+		$this->create();
+	}
+
+
+	/**
+	 * wrapper for update()
+	 *
+	 * @return boolean
+	 */
+	function apply_changes() {
+		if (strtotime($this->created) < self::edit_limit()) {
+			warning(_("This argument may not be updated any longer."));
+			return false;
+		}
+		$this->update(array("title", "content"), "updated=now()");
+	}
+
+
+	/**
 	 * set or update a rating score
 	 *
 	 * @param integer $score
+	 * @return boolean
 	 */
 	function set_rating($score) {
+		if ($this->removed) {
+			warning(_("The argument has been removed."));
+			return false;
+		}
 		if ($score > self::rating_score_max) $score = self::rating_score_max;
 		if ($score < 1) $score = 1;
 		$fields_values = array('argument'=>$this->id, 'member'=>Login::$member->id, 'score'=>$score);
@@ -41,8 +88,14 @@ class Argument extends Relation {
 
 	/**
 	 * reset a rating
+	 *
+	 * @return boolean
 	 */
 	function delete_rating() {
+		if ($this->removed) {
+			warning(_("The argument has been removed."));
+			return false;
+		}
 		DB::delete("ratings", "argument=".intval($this->id)." AND member=".intval(Login::$member->id));
 		$this->update_ratings_cache();
 	}
